# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Opportunity OS CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: organizationId
          schema:
            type: string
        - in: query
          name: tagId
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactListResponse"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get contact by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contact details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateContact
      tags: [contacts]
      summary: Update contact
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete contact
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /opportunities:
    get:
      operationId: listOpportunities
      tags: [opportunities]
      summary: List opportunities
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: pipelineId
          schema:
            type: string
        - in: query
          name: pipelineStageId
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of opportunities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityListResponse"
    post:
      operationId: createOpportunity
      tags: [opportunities]
      summary: Create opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpportunityRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
  /opportunities/{id}:
    get:
      operationId: getOpportunity
      tags: [opportunities]
      summary: Get opportunity by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Opportunity details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityDetail"
    put:
      operationId: updateOpportunity
      tags: [opportunities]
      summary: Update opportunity
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOpportunityRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
    delete:
      operationId: deleteOpportunity
      tags: [opportunities]
      summary: Delete opportunity
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /notes:
    post:
      operationId: createNote
      tags: [notes]
      summary: Create note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    put:
      operationId: updateNote
      tags: [notes]
      summary: Update note
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete note
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Contact:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        organizationId:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        fullName:
          type: string
        title:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        status:
          type: string
        notesText:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        organization:
          $ref: "#/components/schemas/OrganizationSummary"
          nullable: true
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
      required:
        - id
        - workspaceId
        - fullName
        - status
        - createdAt
        - updatedAt
    ContactDetail:
      allOf:
        - $ref: "#/components/schemas/Contact"
        - type: object
          properties:
            activities:
              type: array
              items:
                $ref: "#/components/schemas/Activity"
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/Task"
            notes:
              type: array
              items:
                $ref: "#/components/schemas/Note"
            opportunities:
              type: array
              items:
                $ref: "#/components/schemas/OpportunitySummary"
    ContactListResponse:
      type: object
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - contacts
        - total
        - page
        - limit
    CreateContactRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        title:
          type: string
        department:
          type: string
        email:
          type: string
        phone:
          type: string
        mobile:
          type: string
        linkedinUrl:
          type: string
        source:
          type: string
        organizationId:
          type: string
        notesText:
          type: string
        tagIds:
          type: array
          items:
            type: string
      required:
        - fullName
    UpdateContactRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        title:
          type: string
        department:
          type: string
        email:
          type: string
        phone:
          type: string
        mobile:
          type: string
        linkedinUrl:
          type: string
        source:
          type: string
        status:
          type: string
        organizationId:
          type: string
        notesText:
          type: string
        tagIds:
          type: array
          items:
            type: string
    OrganizationSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        organizationType:
          type: string
        industry:
          type: string
          nullable: true
      required:
        - id
        - name
        - organizationType
...
    Task:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        priority:
          type: string
        status:
          type: string
        contactId:
          type: string
          nullable: true
        organizationId:
          type: string
          nullable: true
        opportunityId:
          type: string
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        contact:
          $ref: "#/components/schemas/Contact"
          nullable: true
        organization:
          $ref: "#/components/schemas/OrganizationSummary"
          nullable: true
      required:
        - id
        - workspaceId
        - title
        - priority
        - status
        - createdAt
        - updatedAt
...
    Activity:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        type:
          type: string
        subject:
          type: string
        description:
          type: string
          nullable: true
        occurredAt:
          type: string
          format: date-time
        contactId:
          type: string
          nullable: true
        organizationId:
          type: string
          nullable: true
        opportunityId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        contact:
          $ref: "#/components/schemas/Contact"
          nullable: true
        organization:
          $ref: "#/components/schemas/OrganizationSummary"
          nullable: true
      required:
        - id
        - workspaceId
        - type
        - subject
        - occurredAt
        - createdAt
...
    OpportunitySummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
        valueEstimate:
          type: number
          nullable: true
        pipelineStageId:
          type: string
      required:
        - id
        - title
        - status
        - pipelineStageId
    Opportunity:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        vertical:
          type: string
        valueEstimate:
          type: number
          nullable: true
        closeDateEstimate:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
        score:
          type: integer
          nullable: true
        source:
          type: string
          nullable: true
        pipelineId:
          type: string
        pipelineStageId:
          type: string
        organizationId:
          type: string
          nullable: true
        primaryContactId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        organization:
          $ref: "#/components/schemas/OrganizationSummary"
          nullable: true
        primaryContact:
          $ref: "#/components/schemas/Contact"
          nullable: true
        pipeline:
          $ref: "#/components/schemas/PipelineSummary"
          nullable: true
        pipelineStage:
          $ref: "#/components/schemas/PipelineStageSummary"
          nullable: true
      required:
        - id
        - workspaceId
        - title
        - vertical
        - status
        - pipelineId
        - pipelineStageId
        - createdAt
        - updatedAt
    OpportunityDetail:
      allOf:
        - $ref: "#/components/schemas/Opportunity"
        - type: object
          properties:
            activities:
              type: array
              items:
                $ref: "#/components/schemas/Activity"
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/Task"
            notes:
              type: array
              items:
                $ref: "#/components/schemas/Note"
            contacts:
              type: array
              items:
                $ref: "#/components/schemas/Contact"
    OpportunityListResponse:
      type: object
      properties:
        opportunities:
          type: array
          items:
            $ref: "#/components/schemas/Opportunity"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - opportunities
        - total
        - page
        - limit
    CreateOpportunityRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        vertical:
          type: string
          enum: [HEALTHCARE, GOVCON, CONSULTING, PARTNERSHIP]
        valueEstimate:
          type: number
        closeDateEstimate:
          type: string
          format: date-time
        source:
          type: string
        pipelineId:
          type: string
        pipelineStageId:
          type: string
        organizationId:
          type: string
        primaryContactId:
          type: string
      required:
        - title
        - pipelineId
        - pipelineStageId
    UpdateOpportunityRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        vertical:
          type: string
        valueEstimate:
          type: number
        closeDateEstimate:
          type: string
          format: date-time
        status:
          type: string
        source:
          type: string
        pipelineId:
          type: string
        pipelineStageId:
          type: string
        organizationId:
          type: string
          nullable: true
        primaryContactId:
          type: string
          nullable: true
...
    PipelineSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name
...
    PipelineStageSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        stageOrder:
          type: integer
      required:
        - id
        - name
        - stageOrder
...
    Note:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        content:
          type: string
        contactId:
          type: string
          nullable: true
        organizationId:
          type: string
          nullable: true
        opportunityId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - content
        - createdAt
        - updatedAt
    CreateNoteRequest:
      type: object
      properties:
        content:
          type: string
        contactId:
          type: string
        organizationId:
          type: string
        opportunityId:
          type: string
      required:
        - content
    UpdateNoteRequest:
      type: object
      properties:
        content:
          type: string
    Tag:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        color:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
      required:
        - id
        - workspaceId
        - name
...