# lib/api-zod/src/generated/types/createOpportunityRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Opportunity OS CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOpportunityRequestVertical } from "./createOpportunityRequestVertical";

export interface CreateOpportunityRequest {
  title: string;
  description?: string;
  vertical?: CreateOpportunityRequestVertical;
  valueEstimate?: number;
  closeDateEstimate?: Date;
  source?: string;
  pipelineId: string;
  pipelineStageId: string;
  organizationId?: string;
  primaryContactId?: string;
}